bot/requirements_frozen.txt
aioamqp==0.11.0
aiohttp==3.3.2
async-timeout==3.0.0
atomicwrites==1.1.5
attrs==18.1.0
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
boto3==1.7.62
botocore==1.10.62
certifi==2018.4.16
chardet==3.0.4
click==6.7
codecov==2.0.15
codespell==1.13.0
colorama==0.3.9
cookies==2.2.1
coverage==4.5.1
coveralls==1.3.0
datadog==0.22.0
decorator==4.3.0
docopt==0.6.2
docutils==0.14
fancycompleter==0.8
flake8==3.5.0
flake8-coding==1.3.0
flake8-copyright==0.2.0
flake8-debugger==3.1.0
flake8-isort==2.5
flake8-mypy==17.8.0
flake8-quotes==1.0.0
httpretty==0.9.5
idna==2.7
idna-ssl==1.1.0
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.12.1
jmespath==0.9.3
Logbook==1.4.0
mccabe==0.6.1
mohawk==0.3.4
more-itertools==4.2.0
mozdef-client==1.0.11
mozilla-cli-common==1.0.0
multidict==4.3.1
mypy==0.620
parsepatch==0.1.1
parso==0.3.1
pathspec==0.5.6
pdbpp==0.9.2
pexpect==4.6.0
pickleshare==0.7.4
pluggy==0.6.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
py==1.5.4
pyasn1==0.3.4
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pytest==3.6.3
pytest-cov==2.5.1
python-dateutil==2.6.1
python-hglib==2.6.1
pytz==2018.5
PyYAML==3.13
raven==6.9.0
RBTools==1.0
requests==2.19.1
requests-futures==0.9.7
responses==0.9.0
s3transfer==0.1.13
simplegeneric==0.8.1
simplejson==3.16.0
six==1.10.0
slugid==1.0.7
structlog==18.1.0
taskcluster==3.0.2
testfixtures==6.2.0
texttable==1.4.0
tqdm==4.23.4
traitlets==4.3.2
typed-ast==1.1.0
urllib3==1.23
wcwidth==0.1.7
wmctrl==0.3
yamllint==1.11.1
yarl==1.2.6
